+++ Gem/Code/Source/AutoGen/NetworkMatchComponent.AutoComponent.xml
@@ -5,7 +5,7 @@
     Namespace="MultiplayerSample"
     OverrideComponent="true"
     OverrideController="true"
-	OverrideInclude="Source/Components/Multiplayer/NetworkMatchComponent.h"
+	OverrideInclude="Source/Components/NetworkMatchComponent.h"
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
 
 	<ComponentRelation Constraint="Required" HasController="true" Name="MatchPlayerCoinsComponent" Namespace="MultiplayerSample" Include="Components/Multiplayer/MatchPlayerCoinsComponent.h" />
